auth and meal repos fixes

+++ app/src/main/java/dev/calorai/mobile/features/meal/data/api/MealApi.kt
@@ -14,7 +14,7 @@ interface MealApi {
 
     @GET("daily-meal/{userId}")
     suspend fun getDailyMeal(
-        @Path("userId") userId: Int,
+        @Path("userId") userId: Long,
         @Query("date") date: String? = null
     ): Response<GetDailyMealResponse>
